ingestion_service/requirements.txt
pandas
requests
python-dotenv
supabase
kagglehub